Småfix
Initierar spelarna i en funktion, gjorde WIDTH mer precis så "glipan" på höger sida försvann. Kommentar om flaggan i testexplosion. Initierar alla spelares invulnerability som false så att inte timern startar 4 ggr första gången nån dör.

+++ Bowlerman/bomb.c
@@ -222,7 +222,7 @@ int correctBowlingBallPosy(int i)
     int offset = WIDTH / 39;
     int j=0, k=0;
     k=(i-WIDTH/297)/size;
-    j=(i-WIDTH/297)%size;
+    j=(int)(i-WIDTH/297)%size;
     if(j<35){
         return (k*size)+offset;      
     }
